+++ src/rpc/blockchain.cpp
@@ -8,9 +8,11 @@
 #include "base58.h"
 #include "checkpoints.h"
 #include "clientversion.h"
+#include "consensus/validation.h"
 #include "main.h"
 #include "policy/feerate.h"
 #include "policy/policy.h"
+#include "primitives/transaction.h"
 #include "rpc/server.h"
 #include "sync.h"
 #include "txdb.h"
