+++ src/txmempool.cpp
@@ -8,6 +8,7 @@
 
 #include "consensus/consensus.h"
 #include "consensus/tx_verify.h"
+#include "consensus/validation.h"
 #include "clientversion.h"
 #include "main.h"
 #include "streams.h"
