+++ src/walletdb.cpp
@@ -8,6 +8,7 @@
 #include "walletdb.h"
 
 #include "consensus/tx_verify.h"
+#include "consensus/validation.h"
 #include "base58.h"
 #include "protocol.h"
 #include "serialize.h"
